+++ Changelog.md
@@ -1,6 +1,7 @@
 ## 0.0.46
 + Added support for converting compressed textures to images with the `MunaUnity.ToImage` extension method.
 + Added `MunaUnity.ToAudioClip` extension method to create an `AudioClip` from an OpenAI speech `BinaryData`.
++ Fixed `DllNotFoundException` when making predictions on macOS in Unity 2022.3 LTS (#14).
 
 ## 0.0.45
 + Added `muna.Beta.OpenAI` property for using predictors with our mock OpenAI client.